requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
discord.py==1.3.4
idna==2.10
multidict==4.7.6
numpy==1.19.1
pandas==1.1.0
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
typing-extensions==3.7.4.2
urllib3==1.25.10
websockets==8.1
yarl==1.5.0
youtube-dl==2020.7.28
